devtools/server/tests/mochitest/test_Debugger.Script.prototype.global.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=958646

Debugger.Script.prototype.global should return innerize globals, not WindowProxies.
-->
<head>
  <meta charset="utf-8">
  <title>Debugger.Script.prototype.global should return inner windows</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script>
"use strict";

const {addDebuggerToGlobal} = ChromeUtils.import("resource://gre/modules/jsdebugger.jsm");
addDebuggerToGlobal(this);

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  const iframe = document.createElement("iframe");
  iframe.src = "data:text/html,<script>function glorp() { }<\/script>";
  iframe.onload = firstOnLoadHandler;
  document.body.appendChild(iframe);

  function firstOnLoadHandler() {
    const dbg = new Debugger();
    const iframeDO = dbg.addDebuggee(iframe.contentWindow);

    // For sanity: check that the debuggee global is the inner window,
    // and that the outer window gets a distinct D.O.
    const iframeWindowProxyDO = iframeDO.makeDebuggeeValue(iframe.contentWindow);
    ok(iframeDO !== iframeWindowProxyDO);

    // The real test: Debugger.Script.prototype.global returns inner windows.
    ok(iframeDO.getOwnPropertyDescriptor("glorp").value.script.global === iframeDO);

    SimpleTest.finish();
  }
};
</script>
</pre>
</body>
</html>
